Move vendir config and related objects to sources package
