+++ cli/pkg/kctrl/local/sources/github_release.go
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package init
+package sources
 
 import (
 	"strings"
