+++ cli/pkg/kctrl/local/sources/vendir_config.go
@@ -1,4 +1,4 @@
-package init
+package sources
 
 import (
 	"os"
