+++ cli/pkg/kctrl/local/sources/vendir_runner.go
@@ -1,4 +1,4 @@
-package init
+package sources
 
 import (
 	"bufio"
